normal user cant use admin panel

+++ blog/views.py
@@ -18,38 +18,45 @@ from django.contrib.auth.hashers import check_password
 from django.core.exceptions import ObjectDoesNotExist
 # -------------------- new category --------------------------------------
 def view_all_categories(request):
-	all_cat = Categories.objects.all()
-	context = {'all_categories':all_cat}
-	return render(request,'blog/category_admin.html',context)
+	if request.user.is_superuser:
+		all_cat = Categories.objects.all()
+		context = {'all_categories':all_cat}
+		return render(request,'blog/category_admin.html',context)
+	return HttpResponseRedirect('/blog/login')
 
 def add_category(request):
-	form = Categories_form()
-	if request.method == 'POST':
-		form = Categories_form(request.POST)
-		if form.is_valid():
-			form.save()
-			return HttpResponseRedirect('/blog/category/all')
-	context = {'category_form':form}
-	return render(request, 'blog/Categories_form.html',context )	
+	if request.user.is_superuser:
+		form = Categories_form()
+		if request.method == 'POST':
+			form = Categories_form(request.POST)
+			if form.is_valid():
+				form.save()
+				return HttpResponseRedirect('/blog/category/all')
+		context = {'category_form':form}
+		return render(request, 'blog/Categories_form.html',context )	
+	return HttpResponseRedirect('/blog/login')
 
 # -------------------- edit category --------------------------------------
 def edit_category(request , slug):
-	cat = Categories.objects.get(slug = slug)
-	form = Categories_form(instance = cat)
-	if request.method == 'POST':
-		form = Categories_form(request.POST, instance = cat)
-		if form.is_valid():
-			form.save()
-			return HttpResponseRedirect('/blog/category/all')
-	context = {'category_form': form}	
-	return render(request, 'blog/Categories_form.html',context)
-
+	if request.user.is_superuser:
+		cat = Categories.objects.get(slug = slug)
+		form = Categories_form(instance = cat)
+		if request.method == 'POST':
+			form = Categories_form(request.POST, instance = cat)
+			if form.is_valid():
+				form.save()
+				return HttpResponseRedirect('/blog/category/all')
+		context = {'category_form': form}	
+		return render(request, 'blog/Categories_form.html',context)
+	return HttpResponseRedirect('/blog/login')
 # -------------------- delete category --------------------------------------
 
 def delete_category(request , slug):
-	cat = Categories.objects.get(slug = slug)
-	cat.delete()
-	return HttpResponseRedirect('/blog/category/all')
+	if request.user.is_superuser:
+		cat = Categories.objects.get(slug = slug)
+		cat.delete()
+		return HttpResponseRedirect('/blog/category/all')
+	return HttpResponseRedirect('/blog/login')
 
 # -------------------- each category_posts -----------------------------------
 
@@ -126,41 +133,48 @@ def view_all_comments(request):
 
 
 def edit_post(request,slug):
-	post=get_object_or_404(Posts,slug=slug)
-	#prepopulated_fields = {'slug':('post_title',)}
-	if request.method == "POST":
-		form=Post_Form(request.POST,request.FILES,instance=post)
-		if form.is_valid():
-			post=form.save(commit=False)
-			#post.auther=request.user
-			post.publish_date=timezone.now()
-			
-			post.save()
-			return HttpResponseRedirect('/blog/admin/posts')
-			#return redirect('view_admin_posts')
-	form=Post_Form(instance=post)
-	context={'post_form':form}
-	return render(request,'blog/postform.html',context)
+	if request.user.is_superuser:
+		post=get_object_or_404(Posts,slug=slug)
+		#prepopulated_fields = {'slug':('post_title',)}
+		if request.method == "POST":
+			form=Post_Form(request.POST,request.FILES,instance=post)
+			if form.is_valid():
+				post=form.save(commit=False)
+				#post.auther=request.user
+				post.publish_date=timezone.now()
+				
+				post.save()
+				return HttpResponseRedirect('/blog/admin/posts')
+				#return redirect('view_admin_posts')
+		form=Post_Form(instance=post)
+		context={'post_form':form}
+		return render(request,'blog/postform.html',context)
+
+	return HttpResponseRedirect('/blog/login')
 
 
 def add_post(request):
-	form=Post_Form()
-	if request.method == 'POST':
-		form=Post_Form(request.POST,request.FILES)
-		if form.is_valid():
-			post=form.save(commit=False)
-			post.author=request.user
-			post.save()
-			return HttpResponseRedirect('/blog/admin/posts')
-	context={'post_form':form}
-	return render(request,'blog/postform.html',context)
+	if request.user.is_superuser:
+		form=Post_Form()
+		if request.method == 'POST':
+			form=Post_Form(request.POST,request.FILES)
+			if form.is_valid():
+				post=form.save(commit=False)
+				post.author=request.user
+				post.save()
+				return HttpResponseRedirect('/blog/admin/posts')
+		context={'post_form':form}
+		return render(request,'blog/postform.html',context)
+	return HttpResponseRedirect('/blog/login')
 
 
 def delete_post(request,slug):
-	post=Posts.objects.get(slug=slug)
-	Comment_Section.objects.filter(comment_post=post.id).delete()# de iterator
-	post.delete()
-	return HttpResponseRedirect('/blog/admin/posts')
+	if request.user.is_superuser:
+		post=Posts.objects.get(slug=slug)
+		Comment_Section.objects.filter(comment_post=post.id).delete()# de iterator
+		post.delete()
+		return HttpResponseRedirect('/blog/admin/posts')
+	return HttpResponseRedirect('/blog/login')
 
 
 
@@ -196,61 +210,73 @@ def unsubscribe(request,slug):
 #-----------------Administration---------------------------------------
 
 def view_all_users (request):
-	users=User.objects.all()
-	context={'all_users': users}
-	return render (request,'blog/All_Users.html',context)
+	if request.user.is_superuser:
+		users=User.objects.all()
+		context={'all_users': users}
+		return render (request,'blog/All_Users.html',context)
+	return HttpResponseRedirect('/blog/login')
 
-def block_users (request,user_id):	
-	usr=User.objects.get(id=user_id)
-	if usr.is_active is True:
-		usr.is_active = False
-		usr.save()
-		return HttpResponseRedirect('/blog/users')
-	else:
-		usr.is_active = True
-		usr.save()
-		return HttpResponseRedirect('/blog/users')	
+def block_users (request,user_id):
+	if request.user.is_superuser:	
+		usr=User.objects.get(id=user_id)
+		if usr.is_active is True:
+			usr.is_active = False
+			usr.save()
+			return HttpResponseRedirect('/blog/users')
+		else:
+			usr.is_active = True
+			usr.save()
+			return HttpResponseRedirect('/blog/users')
+	return HttpResponseRedirect('/blog/login')	
 
 def delete_user (request,user_id):
-	num_admins=User.objects.filter(is_superuser=True).count()
-	print num_admins
-	if num_admins == 1:
-		print"You can't delete yourself"
-		return HttpResponseRedirect('/blog/users')
-	else:	
-		user=User.objects.get(id=user_id)
-		user.delete()
-		return HttpResponseRedirect('/blog/users')
+	if request.user.is_superuser:
+		num_admins=User.objects.filter(is_superuser=True).count()
+		if num_admins == 1:
+			print"You can't delete yourself"
+			return HttpResponseRedirect('/blog/users')
+		else:	
+			user=User.objects.get(id=user_id)
+			user.delete()
+			return HttpResponseRedirect('/blog/users')
+
+	return HttpResponseRedirect('/blog/login')
 		
 
 def promote_user(request,user_id):
-	user=User.objects.get(id=user_id)
-	if user.is_superuser:
-		user.is_superuser = False
-		user.save()
-
-		return HttpResponseRedirect('/blog/users')
-
-	else:
-		user.is_superuser= True
-	user.save()
-	return HttpResponseRedirect('/blog/users')
+	if request.user.is_superuser:
+		user=User.objects.get(id=user_id)
+		if user.is_superuser:
+				user.is_superuser = False
+				user.save()
+				return HttpResponseRedirect('/blog/users')
+	
+		else:
+			user.is_superuser= True
+			user.save()
+			return HttpResponseRedirect('/blog/users')
+	return HttpResponseRedirect('/blog/login')
 
 def all_inappr(request):
-	all_inappr=Inappropriate_words.objects.all()
-	context={'all_inappr': all_inappr}
-
-	return render (request,'blog/all_inappr.html',context)
+	if request.user.is_superuser:
+		all_inappr=Inappropriate_words.objects.all()
+		context={'all_inappr': all_inappr}
+	
+		return render (request,'blog/all_inappr.html',context)
+	return HttpResponseRedirect('/blog/login')
 
 def new_inappr(request):
-	form = Inappr_Form()
-	if request.method == "POST":
-		form = Inappr_Form(request.POST)
-		if form.is_valid():
-			form.save()
-			return HttpResponseRedirect('/blog/inappr')
-	context={'inappr_form':form}
-	return render(request, 'blog/new_inappr.html',context)
+	if request.user.is_superuser:
+		form = Inappr_Form()
+		if request.method == "POST":
+			form = Inappr_Form(request.POST)
+			if form.is_valid():
+				form.save()
+				return HttpResponseRedirect('/blog/inappr')
+		context={'inappr_form':form}
+		return render(request, 'blog/new_inappr.html',context)
+
+	return HttpResponseRedirect('/blog/login')
 
 
 
@@ -261,33 +287,44 @@ class RegisterView(generic.CreateView):
     template_name = "registration/signup.html"
 
 def dashboard(request):
-	num_cat = Categories.objects.all().count()
-	num_post=Posts.objects.all().count()
-	num_words=Inappropriate_words.objects.all().count()
-	num_user=User.objects.all().count()
-	
-	context = {'num_cat':num_cat ,'num_post':num_post ,'num_words':num_words,'num_user':num_user}
-	return render(request,'blog/dashboard.html',context)
+	if request.user.is_superuser:
+		num_cat = Categories.objects.all().count()
+		num_post=Posts.objects.all().count()
+		num_words=Inappropriate_words.objects.all().count()
+		num_user=User.objects.all().count()
+		context = {'num_cat':num_cat ,'num_post':num_post ,'num_words':num_words,'num_user':num_user}
+		return render(request,'blog/dashboard.html',context)
+
+	return HttpResponseRedirect('/blog/login')
 
 def view_admin_posts(request):
-	all_posts= Posts.objects.all()
-	paginator=Paginator(all_posts,6)
-	page =request.GET.get('page')
-	try:
-		context =paginator.page(page)
-	except PageNotAnInteger:
-		context =paginator.page(1)
-	except EmptyPage:
-		context =paginator.page(paginator.num_pages)
-	return render(request,'blog/posts_admin.html',{'page_counter':context})
+	if request.user.is_superuser:
+		all_posts= Posts.objects.all()
+		paginator=Paginator(all_posts,6)
+		page =request.GET.get('page')
+		try:
+			context =paginator.page(page)
+		except PageNotAnInteger:
+			context =paginator.page(1)
+		except EmptyPage:
+			context =paginator.page(paginator.num_pages)
+		return render(request,'blog/posts_admin.html',{'page_counter':context})
+
+	return HttpResponseRedirect('/blog/login')
 
 	#------------------- authentication on login -----------------------------
+
 def my_login_view(request):
+    if request.user.is_authenticated:
+        return HttpResponseRedirect('/blog')
+
+
+
     form = forms.LoginForm()
     errMsg = ""
-    form = forms.LoginForm(request.POST)
-    print "Form", form
+    
     if request.method == "POST":
+        form = forms.LoginForm(request.POST)
         if not form.is_valid() or form.is_valid():
             username = form.cleaned_data["username"]
             password = form.cleaned_data["password"]
